pyproject.toml
[tool.poetry]
name = "habbit-tracker"
version = "0.1.0"
description = ""
authors = ["Said Magomedov <saidmagomedov037@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.3.0"
django = "^5.2.4"
djangorestframework = "^3.16.0"
redis = "^6.2.0"
djangorestframework-simplejwt = "^5.5.1"
python-dotenv = "^1.1.1"
psycopg2 = "^2.9.10"
celery = "^5.5.3"
django-celery-beat = "^2.8.1"
black = "^25.1.0"
drf-yasg = "^1.21.10"
coverage = "^7.10.2"
requests = "^2.32.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
